refactor: improve esm
