+++ main.js
@@ -1,0 +1,5 @@
+/* eslint no-global-assign: 0 */
+
+require = require('esm')(module);
+
+module.exports = require('./generators/app');
